+++ SmartHouseAppServer/SmartHouseApp.Share/ViewModel/SizeViewModel.cs
@@ -8,7 +8,7 @@ namespace SmartHouseApp.Share.ViewModel
 {
     public class SizeViewModel
     {
-        public virtual int X { get; set; }
-        public virtual int Y { get; set; }
+        public virtual double X { get; set; }
+        public virtual double Y { get; set; }
     }
 }
